+++ .eslintrc
@@ -34,10 +34,6 @@
     "react/no-unused-prop-types": 0,
     "import/export": 0,
     "no-unused-vars": 1,
-    "react/jsx-uses-react": "off",
-    "react/react-in-jsx-scope": "off",
-    "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "warn",
     "prettier/prettier": "error",
     "import/no-anonymous-default-export": "off"
   }
