+++ package.json
@@ -16,7 +16,7 @@
     "src/loader/index.js",
     "src/loader/loader-helpers.js",
     "src/rcc/rcc-core.js",
-    "src/rcc/rcc-core.tsx"
+    "src/rcc/rcc-core.ts"
   ],
   "exports": {
     ".": "./dist/index.js",
@@ -48,11 +48,11 @@
     "@types/jest": "^25.1.4",
     "@types/mocha": "^9.1.1",
     "@types/node": "^18.6.4",
-    "@typescript-eslint/eslint-plugin": "^2.26.0",
-    "@typescript-eslint/parser": "^2.26.0",
-    "babel-eslint": "^10.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "babel-eslint": "^10.1.0",
     "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
+    "eslint": "^8.24.0",
     "eslint-config-prettier": "^6.7.0",
     "eslint-config-standard": "^14.1.0",
     "eslint-config-standard-react": "^9.2.0",
@@ -62,7 +62,6 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
     "gh-pages": "^2.2.0",
-    "microbundle": "^0.15.1",
     "microbundle-crl": "^0.13.10",
     "mocha": "^10.0.0",
     "npm-run-all": "^4.1.5",
@@ -70,7 +69,7 @@
     "prettier": "^2.0.4",
     "sass": "^1.54.3",
     "tsc": "^2.0.4",
-    "typescript": "^3.9.10"
+    "typescript": "^4.8.4"
   },
   "keywords": [
     "react",
