+++ src/rcc/helpers.ts
@@ -6,21 +6,11 @@ export const toPascalCase = (str: string) =>
 export const toCamelCase = (str: string) =>
   str.replace(/-[a-z0-9]/g, (match) => match.toUpperCase().replace('-', ''))
 
-type ParseClassKey<K extends string> = K extends `${infer P1}-${infer P2}`
-  ? ParseClassKey<`${P1}${Capitalize<P2>}`>
-  : K
-
 export const classNamesMapping = <
   T extends Record<string, string | { [K: string]: string }>
 >(
   classnames?: T
 ) => {
-  type Params = {
-    [K in keyof T as ParseClassKey<K & string>]?: T[K] extends string
-      ? boolean
-      : keyof T[K]
-  }
-
   const valuesMapping = Object.entries(classnames || {}).reduce(
     (prev, [key, value]) => {
       const camelKey = toCamelCase(key)
@@ -29,7 +19,9 @@ export const classNamesMapping = <
     },
     {}
   )
-  const getClassName = (params?: Params & { className?: string }) => {
+  const getClassName = (
+    params?: { [K: string]: any } & { className?: string }
+  ) => {
     const { className, ...rest } = params || {}
     return Object.entries(rest).reduce((prev, [key, value]) => {
       const cn = valuesMapping[key]?.[String(value)]
